pyproject.toml
[tool.poetry]
name = "callmeuncle"
version = "0.1.0"
description = ""
authors = ["Arush Kurundodi <arushkurundodi@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
google-genai = "^0.3.0"
asyncio = "^3.4.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
